Modified level of some debug output. Moved CoreTypeMapperTest to the correct folder.
