+++ bundles/binding/org.openhab.binding.knx/src/main/java/org/openhab/binding/knx/internal/connection/KNXConnectionListener.java
@@ -10,7 +10,7 @@ package org.openhab.binding.knx.internal.connection;
 
 /**
  * This interface should be implemented by classes that need to be informed when
- * a KNX connection has been established.
+ * a KNX connection has been established or was lost.
  * 
  * @author Andrea Giacosi
  * @author Kai Kreuzer
@@ -19,7 +19,7 @@ package org.openhab.binding.knx.internal.connection;
 public interface KNXConnectionListener {
 	
 	/**
-	 * The callback method that is used to notify listeners about a successfull KNX connection
+	 * The callback method that is used to notify listeners about a successful KNX connection
 	 */
 	void connectionEstablished();
